fix get status
